refactor(web): genericise shared stacking form components
