+++ apps/web/app/features/stacking/utils/stacking-network-types.ts
@@ -1,0 +1,1 @@
+export type NetworkMode = 'mainnet' | 'testnet' | 'devnet';
